+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Fixed
+- Japanese in english locales for Gotoaddress, addclass dialog
+
 ## [14.5.2] - 2021-11-20
 ### Fixed
 - AS1/2 handle declaration of registers in certain cases