+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Reload menu disabled when do SWF selected
 - AS2 - Do not detect classes inside functions
 - AS1/2 - Slash syntax colon vs ternar operator collision
+- AS1/2 - Allow nonstandard identifiers in object literal
 
 ## [14.5.2] - 2021-11-20
 ### Fixed