+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - AS1/2 - Slash syntax colon vs ternar operator collision
 - AS1/2 - Allow nonstandard identifiers in object literal
 - AS1/2 - Allow globalfunc names as variable identifiers
+- AS1/2 - Registers in for..in clause, proper define
 
 ## [14.5.2] - 2021-11-20
 ### Fixed