+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [14.6.0] - 2021-11-22
 ### Added
 - Information message before importing scripts, text, XML, Symbol-Class